[MOV] Moved fields Freight, Insurance and Other Costs to l10n_br_sale because already exist there.
